requirements.txt
qrcode
requests
beautifulsoup4
sumy
pillow
pygame
opencv-python
pyzbar
numpy
logging